initial restructuring of agents
simplify agents: just use `give-recommendations.ts` agent for assets|vulnerabilities.
