--- src/features/chat/viper-agent/agents/explain-vulnerability.ts
@@ -1,34 +1,0 @@
-import {
-  createAgent,
-  type NetworkRun,
-  type StateData,
-} from "@inngest/agent-kit";
-import { type UserRole, VULNERABILITY_ROLE_INSTRUCTIONS } from "../../utils";
-import { DEFAULT_CHAT_MODEL } from "../constants";
-import { getVulnerabilityData } from "../tools/get-vulnerability-data";
-
-const MODEL = DEFAULT_CHAT_MODEL;
-
-const buildSystemPrompt = (
-  network: NetworkRun<StateData> | undefined,
-): string => {
-  const data = network?.state.data as { userRole?: UserRole } | undefined;
-  const role: UserRole = data?.userRole ?? "hospital administration";
-  const roleInstructions = VULNERABILITY_ROLE_INSTRUCTIONS[role];
-
-  return `You are an AI assistant helping analyze a vulnerability in the Viper hospital vulnerability management platform.
-Always call the get_vulnerability_data tool first to retrieve full vulnerability details before answering any question.
-Use the returned data to give accurate, grounded answers — do not invent numbers or details.
-
-${roleInstructions}`;
-};
-
-export const createExplainVulnerabilityAgent = () =>
-  createAgent({
-    name: "Viper Vulnerability Explainer",
-    description:
-      "Explains a hospital vulnerability's risk, affected assets, and remediation options.",
-    system: ({ network }) => buildSystemPrompt(network),
-    tools: [getVulnerabilityData],
-    model: MODEL,
-  });
