--- src/features/chat/viper-agent/tools/get-vulnerability-data.ts
@@ -1,46 +1,0 @@
-import "server-only";
-import { createTool } from "@inngest/agent-kit";
-import { z } from "zod";
-
-// TODO: made this a tool so we can expand it in the future
-export const getVulnerabilityData = createTool({
-  name: "get_vulnerability_data",
-  description:
-    "Returns detailed information about the vulnerability currently being discussed, including affected assets and remediations. Always call this tool before answering questions about the vulnerability.",
-  parameters: z.object({}),
-  handler: async (_, { network }) => {
-    const vulnerability = network?.state.data.vulnerabilityData;
-    if (!vulnerability) return "No vulnerability data available.";
-
-    const issues = Array.isArray(vulnerability.issues)
-      ? vulnerability.issues
-      : [];
-    const assets = issues.map((i: { asset: unknown }) => i.asset);
-    const remediations = Array.isArray(vulnerability.remediations)
-      ? vulnerability.remediations
-      : [];
-
-    // TODO: eventually just reuse vulnerabiltyToMarkdown in get-recommendations-context
-    return [
-      `## Vulnerability: ${vulnerability.cveId ?? vulnerability.id}`,
-      "",
-      `- **CVE ID**: ${vulnerability.cveId ?? "N/A"}`,
-      `- **CVSS Score**: ${vulnerability.cvssScore ?? "N/A"}`,
-      `- **EPSS Score**: ${vulnerability.epss != null ? `${(vulnerability.epss * 100).toFixed(2)}%` : "N/A"}`,
-      `- **Severity**: ${vulnerability.severity ?? "N/A"}`,
-      `- **Created**: ${vulnerability.createdAt ?? "N/A"}`,
-      "",
-      "## Vulnerability Details",
-      "",
-      JSON.stringify(vulnerability, null, 2),
-      "",
-      "## Affected Assets",
-      "",
-      JSON.stringify(assets, null, 2),
-      "",
-      "## Remediations",
-      "",
-      JSON.stringify(remediations, null, 2),
-    ].join("\n");
-  },
-});
