+++ src/features/vulnerabilities/components/vulnerability-drawer.tsx
@@ -385,7 +385,7 @@ export function PrioritizedVulnerabilityDrawer({
         <SuggestedQuestionsProvider questions={visibleQuestions}>
           <AIChat
             config={{
-              agent: "explainVulnerability",
+              agent: "giveRecommendations",
               vulnerabilityData: vulnerability,
             }}
           />
