+++ src/inngest/functions/chat-agent.ts
@@ -9,8 +9,6 @@ import {
 import { createChannel } from "@/app/api/inngest/realtime";
 import type { NetworkState } from "@/features/chat/types";
 import { createChatAgent } from "@/features/chat/viper-agent/agents/chat-agent";
-import { createExplainAssetAgent } from "@/features/chat/viper-agent/agents/explain-asset";
-import { createExplainVulnerabilityAgent } from "@/features/chat/viper-agent/agents/explain-vulnerability";
 import { createGiveRecommendationsAgent } from "@/features/chat/viper-agent/agents/give-recommendations";
 import { conversationHistoryAdapter } from "@/features/chat/viper-agent/history-adapter";
 import { inngest } from "../client";
@@ -33,12 +31,6 @@ export const chatAgent = inngest.createFunction(
 
     let agent: Agent<StateData>;
     switch (clientState.agent) {
-      case "explainAsset":
-        agent = createExplainAssetAgent();
-        break;
-      case "explainVulnerability":
-        agent = createExplainVulnerabilityAgent();
-        break;
       case "giveRecommendations":
         agent = createGiveRecommendationsAgent();
         break;
